backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
autopep8==2.3.2
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
dotenv-python==0.0.1
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
inflect==7.5.0
more-itertools==10.8.0
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.2
sqlacodegen==4.0.3
SQLAlchemy==2.0.48
starlette==0.52.1
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
